pyproject.toml
[project]
name = "fastapi-prod-backend"
version = "0.1.0"
description = "Production-ready FastAPI backend with SQLAlchemy and Alembic"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "pydantic-settings>=2.12.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-describe>=3.1.0",
]
